feat: add flag for more explanations
Added a flag parameter to the command line script that when set to
true, prints the commit context and message to the console instead of
creating and saving the commit message to the repository. This allows
user to have more information about the commit message and have more
control over it.