feat(gpt3_cost_calculator): calculate the number of tokens needed for the diff
This commit adds the logic needed to calculate the number of tokens
needed to represent a diff in the gpt3_cost_calculator.go file. The
calculation is based on the GPT-2 tokenizer in combination with the
OpenAI GPT-3 API. It also adds a function that can be used to shorten a
string to the required amount of tokens. This change has the potential
to improve the accuracy of the cost calculation for GPT-3.